# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Via-Alias API",
    "description": "API for [Via-Alias](https://github.com/ngundlach/via-alias). This API can be used to create and manage redirects as well as users.\n        <br><br>**Authentication:**<br>\n        Most Endpoints require a Bearer token obtained from `POST /api/auth/login`.\n        Include it in the `Authorization` header as `Bearer <token>`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.11"
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Example:"
    }
  ],
  "paths": {
...
    "/api/admin/redirects/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete redirect via id",
        "description": "Deletes a redirect via its id. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "delete_redirect_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The redirect id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content. Deletes a redirect"
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          },
          "404": {
            "description": "Not Found. Redirect doesn't exist."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/redirects": {
      "get": {
        "tags": [
          "Redirects"
        ],
        "summary": "Get redirects",
        "description": "Returns a list of all redirects owned by the current user. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "get_user_redirects_list",
        "responses": {
          "200": {
            "description": "Ok. Returns list of redirects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectListDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Redirects"
        ],
        "summary": "Create redirect",
        "description": "Creates a new redirect. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "create_redirect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created. Returns list of redirects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectDTO"
                }
              }
            }
          },
          "400": {
            "description": "Alias or url don't match requirements."
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "409": {
            "description": "A redirect with that alias already exists."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/redirects/{alias}": {
      "delete": {
        "tags": [
          "Redirects"
        ],
        "summary": "Delete redirect",
        "description": "Deletes a redirect via its alias. Users can only delete redirects they have created. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "delete_redirect",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "description": "The redirect alias.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content. Redirect deleted successfully."
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Redirects"
        ],
        "summary": "Update redirect",
        "description": "Updates the registered url of a redirect. Changing the alias requires deleting the redirect and recreating it with the desired alias.\n    Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "update_redirect",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "description": "The redirect alias.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUrlDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok. Changed the url of the redirect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authenticated but doesn't have permission."
          },
          "404": {
            "description": "Not Found. Redirect doesn't exist."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/users/info": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get own userdata",
        "description": "Returns userdata about the current user. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "get_user_data",
        "responses": {
          "200": {
            "description": "Ok. Returns userdata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleUserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/users/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Change password",
        "description": "Changes the user-password. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeDataDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok. Password changed successfully."
          },
          "400": {
            "description": "Bad Request. Current password is invalid or wew password doesn't match requirements."
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register new user",
        "description": "Registers a new user. Requires a generated user registration token.",
        "operationId": "register_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created. User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleUserDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Password or username don't match requirements."
          },
          "409": {
            "description": "Conflict. A user with that name already exists."
          }
        },
        "security": []
      }
    },
...
    "/{alias}": {
      "get": {
        "tags": [
          "Redirects"
        ],
        "summary": "Follow redirect",
        "description": "Returns a tomporary redirect response with the registered redirect url in the location header. Supposed to be opened in a browser.",
        "operationId": "follow_redirect",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "description": "The redirect alias.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Temporary Redirect. Follow redirect",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL of the created resource"
              }
            }
          },
          "404": {
            "description": "Not Found. Redirect doesn't exist."
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "PasswordChangeDataDTO": {
        "type": "object",
        "title": "PasswordChangeData",
        "required": [
          "current_pw",
          "new_pw"
        ],
        "properties": {
          "current_pw": {
            "type": "string",
            "examples": [
              "superjedimeister1337"
            ]
          },
          "new_pw": {
            "type": "string",
            "examples": [
              "kfdslfoi32j4ekil32jm"
            ]
          }
        }
      },
...
      "RedirectDTO": {
        "type": "object",
        "title": "SimpleRedirectData",
        "required": [
          "alias",
          "url"
        ],
        "properties": {
          "alias": {
            "type": "string",
            "examples": [
              "gh"
            ]
          },
          "url": {
            "type": "string",
            "examples": [
              "http://www.github.com"
            ]
          }
        }
      },
      "RedirectListDTO": {
        "type": "object",
        "title": "RedirectList",
        "required": [
          "redirects"
        ],
        "properties": {
          "redirects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedirectDTO"
            }
          }
        }
      },
      "SimpleUserDTO": {
        "type": "object",
        "title": "SimpleUserData",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "examples": [
              "b208ae42-2ab8-47b9-9c69-bdd7300e5847"
            ]
          },
          "name": {
            "type": "string",
            "examples": [
              "vader"
            ]
          }
        }
      },
      "UpdateUrlDTO": {
        "type": "object",
        "title": "UpdateUrl",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "examples": [
              "http://my-new-redirect-url.de"
            ]
          }
        }
      },
...
      "UserRegistrationDTO": {
        "type": "object",
        "title": "UserRegistration",
        "required": [
          "name",
          "pw",
          "token"
        ],
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "vader"
            ]
          },
          "pw": {
            "type": "string",
            "examples": [
              "deathstar_pilot_extra_ordinaire1337"
            ]
          },
          "token": {
            "type": "string",
            "examples": [
              "8162f4c7-e32e-45bc-9089-bfd9d024e39e"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}